+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Added
 - Enum can now have `rg`, `ud`, `void`, `pat` has value type
 - Object can have `const` properties (https://github.com/buzz-language/buzz/issues/13). A object with only `const` properties is considered itself `const`. Although we don't do anything yet with that concept. https://github.com/buzz-language/buzz/issues/114 is the objective but it requires being able to build objects and instances at compile time which is not yet possible.
-- `rg.subsetOf`, `rg.intersect`
+- `rg.subsetOf`, `rg.intersect`, `rg.union`
 
 ## Modified